requirements.txt
easydict==1.13
PyYAML==6.0.1
Requests==2.31.0
